pyproject.toml
[project]
name = "nl-crime-dashboard"
version = "0.1.0"
description = "Netherlands crime data pipeline and interactive dashboard"
requires-python = ">=3.10"
dependencies = [
    "cbsodata>=1.3",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "geoalchemy2>=0.14",
    "streamlit>=1.30",
    "folium>=0.15",
    "streamlit-folium>=0.18",
    "plotly>=5.18",
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.2",
]

[tool.ruff]
line-length = 100
target-version = "py310"
